fix: uncomment posthog
