+++ src/utils/featureFlagChecks/static.tsx
@@ -1,0 +1,7 @@
+import { FLAGS } from "./flags";
+
+export function isFeatureFlagEnabledStatic(
+  featureFlagKey: keyof typeof FLAGS,
+): boolean {
+  return FLAGS[featureFlagKey] === "true";
+}
